pyproject.toml
[tool.poetry]
name = "trust-network"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.1"
pydantic = "^1.10.7"
uvicorn = "^0.22.0"
asyncpg = "^0.27.0"
tortoise-orm = "^0.19.3"
pytest = "^7.3.1"
asgi-lifespan = "^2.1.0"
httpx = "^0.24.0"
email-validator = "^2.0.0.post2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
